Relocates uniter access control method generation to a new module for clarity.
